// front/src/providers/api-client.generated/model/customerDto.ts
/**
 * Morgan And Mallet CRM
 * Morgan And Mallet CRM API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AddressDto } from './addressDto';
import { AppValueDto } from './appValueDto';


export interface CustomerDto { 
    id?: string;
    firstName?: string;
    lastName?: string;
    email?: string;
    phone?: string;
    addresses?: Array<AddressDto>;
    creationDate?: Date;
    modifDate?: Date;
    dateOfContact?: Date;
    isPrivatePerson?: boolean;
    isCompany?: boolean;
    companyName?: string;
    disabled?: boolean;
    contactFullName?: string;
    customerFunctionId?: string;
    customerFunction?: AppValueDto;
}